move GL init and info dump into GL renderer
