+++ src/graphics/Graphics.cpp
@@ -4,7 +4,6 @@
 #include "Graphics.h"
 #include "FileSystem.h"
 #include "Material.h"
-#include "opengl/gl_core_3_x.h"
 #include "opengl/RendererGL.h"
 #include "dummy/RendererDummy.h"
 #include "OS.h"
@@ -46,133 +45,6 @@ float GetFovFactor()
 	return g_fovFactor;
 }
 
-static const char *gl_error_to_string(GLenum err)
-{
-	switch (err) {
-		case GL_NO_ERROR: return "(no error)";
-		case GL_INVALID_ENUM: return "invalid enum";
-		case GL_INVALID_VALUE: return "invalid value";
-		case GL_INVALID_OPERATION: return "invalid operation";
-		case GL_INVALID_FRAMEBUFFER_OPERATION: return "invalid framebuffer operation";
-		case GL_OUT_OF_MEMORY: return "out of memory";
-		default: return "(unknown error)";
-	}
-}
-
-static void dump_and_clear_opengl_errors(std::ostream &out, GLenum first_error = GL_NO_ERROR)
-{
-	GLenum err = ((first_error == GL_NO_ERROR) ? glGetError() : first_error);
-	if (err != GL_NO_ERROR) {
-		out << "errors: ";
-		do {
-			out << gl_error_to_string(err) << " ";
-			err = glGetError();
-		} while (err != GL_NO_ERROR);
-		out << std::endl;
-	}
-}
-
-static void dump_opengl_value(std::ostream &out, const char *name, GLenum id, int num_elems)
-{
-	assert(num_elems > 0 && num_elems <= 4);
-	assert(name);
-
-	GLdouble e[4];
-	glGetDoublev(id, e);
-
-	GLenum err = glGetError();
-	if (err == GL_NO_ERROR) {
-		out << name << " = " << e[0];
-		for (int i = 1; i < num_elems; ++i)
-			out << ", " << e[i];
-		out << "\n";
-	} else {
-		while (err != GL_NO_ERROR) {
-			if (err == GL_INVALID_ENUM) { out << name << " -- not supported\n"; }
-			else { out << name << " -- unexpected error (" << err << ") retrieving value\n"; }
-			err = glGetError();
-		}
-	}
-}
-
-static void write_opengl_info(std::ostream &out)
-{
-	out << "OpenGL version " << glGetString(GL_VERSION);
-	out << ", running on " << glGetString(GL_VENDOR);
-	out << " " << glGetString(GL_RENDERER) << "\n";
-
-	out << "Available extensions:" << "\n";
-	{
-		out << "Shading language version: " <<  glGetString(GL_SHADING_LANGUAGE_VERSION) << "\n";
-		GLint numext = 0;
-		glGetIntegerv(GL_NUM_EXTENSIONS, &numext);
-		for (int i = 0; i < numext; ++i) {
-			out << "  " << glGetStringi(GL_EXTENSIONS, i) << "\n";
-		}
-	}
-
-	out << "\nImplementation Limits:\n";
-
-	// first, clear all OpenGL error flags
-	dump_and_clear_opengl_errors(out);
-
-#define DUMP_GL_VALUE(name) dump_opengl_value(out, #name, name, 1)
-#define DUMP_GL_VALUE2(name) dump_opengl_value(out, #name, name, 2)
-
-	DUMP_GL_VALUE(GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS);
-	DUMP_GL_VALUE(GL_MAX_CUBE_MAP_TEXTURE_SIZE);
-	DUMP_GL_VALUE(GL_MAX_DRAW_BUFFERS);
-	DUMP_GL_VALUE(GL_MAX_ELEMENTS_INDICES);
-	DUMP_GL_VALUE(GL_MAX_ELEMENTS_VERTICES);
-	DUMP_GL_VALUE(GL_MAX_FRAGMENT_UNIFORM_COMPONENTS);
-	DUMP_GL_VALUE(GL_MAX_TEXTURE_IMAGE_UNITS);
-	DUMP_GL_VALUE(GL_MAX_TEXTURE_LOD_BIAS);
-	DUMP_GL_VALUE(GL_MAX_TEXTURE_SIZE);
-	DUMP_GL_VALUE(GL_MAX_VERTEX_ATTRIBS);
-	DUMP_GL_VALUE(GL_MAX_VERTEX_TEXTURE_IMAGE_UNITS);
-	DUMP_GL_VALUE(GL_MAX_VERTEX_UNIFORM_COMPONENTS);
-	DUMP_GL_VALUE(GL_NUM_COMPRESSED_TEXTURE_FORMATS);
-	DUMP_GL_VALUE(GL_SAMPLE_BUFFERS);
-	DUMP_GL_VALUE(GL_SAMPLES);
-	DUMP_GL_VALUE2(GL_ALIASED_LINE_WIDTH_RANGE);
-	DUMP_GL_VALUE2(GL_MAX_VIEWPORT_DIMS);
-	DUMP_GL_VALUE2(GL_SMOOTH_LINE_WIDTH_RANGE);
-	DUMP_GL_VALUE2(GL_SMOOTH_POINT_SIZE_RANGE);
-
-#undef DUMP_GL_VALUE
-#undef DUMP_GL_VALUE2
-
-	// enumerate compressed texture formats
-	{
-		dump_and_clear_opengl_errors(out);
-		out << "\nCompressed texture formats:\n";
-
-		GLint nformats;
-		GLint formats[128]; // XXX 128 should be enough, right?
-
-		glGetIntegerv(GL_NUM_COMPRESSED_TEXTURE_FORMATS, &nformats);
-		GLenum err = glGetError();
-		if (err != GL_NO_ERROR) {
-			out << "Get NUM_COMPRESSED_TEXTURE_FORMATS failed\n";
-			dump_and_clear_opengl_errors(out, err);
-		} else {
-			assert(nformats >= 0 && nformats < int(COUNTOF(formats)));
-			glGetIntegerv(GL_COMPRESSED_TEXTURE_FORMATS, formats);
-			err = glGetError();
-			if (err != GL_NO_ERROR) {
-				out << "Get COMPRESSED_TEXTURE_FORMATS failed\n";
-				dump_and_clear_opengl_errors(out, err);
-			} else {
-				for (int i = 0; i < nformats; ++i) {
-					out << stringf("  %0{x#}\n", unsigned(formats[i]));
-				}
-			}
-		}
-	}
-	// one last time
-	dump_and_clear_opengl_errors(out);
-}
-
 Renderer* Init(Settings vs)
 {
 	assert(!initted);
@@ -191,26 +63,6 @@ Renderer* Init(Settings vs)
 	width = window->GetWidth();
 	height = window->GetHeight();
 
-	const int didLoad = ogl_LoadFunctions();
-	if (!didLoad)
-		Error("glLoadGen failed to load functions.\n");
-
-	{
-		std::ostringstream buf;
-		write_opengl_info(buf);
-
-		FILE *f = FileSystem::userFiles.OpenWriteStream("opengl.txt", FileSystem::FileSourceFS::WRITE_TEXT);
-		if (!f)
-			Output("Could not open 'opengl.txt'\n");
-		const std::string &s = buf.str();
-		fwrite(s.c_str(), 1, s.size(), f);
-		fclose(f);
-	}
-
-	if (ogl_ext_EXT_texture_compression_s3tc == ogl_LOAD_FAILED) {
-		Error("OpenGL extension GL_EXT_texture_compression_s3tc not supported.\nPioneer can not run on your graphics card as it does not support compressed (DXTn/S3TC) format textures.");
-	}
-
 	// We deliberately ignore the value from GL_NUM_COMPRESSED_TEXTURE_FORMATS, because some drivers
 	// choose not to list any formats (despite supporting texture compression). See issue #3132.
 	// This is (probably) allowed by the spec, which states that only formats which are "suitable
@@ -224,8 +76,17 @@ Renderer* Init(Settings vs)
 
 	Output("Initialized %s\n", renderer->GetName());
 
-	std::ostringstream dummy;
-	dump_and_clear_opengl_errors(dummy);
+	{
+		std::ostringstream buf;
+		renderer->WriteRendererInfo(buf);
+
+		FILE *f = FileSystem::userFiles.OpenWriteStream("opengl.txt", FileSystem::FileSourceFS::WRITE_TEXT);
+		if (!f)
+			Output("Could not open 'opengl.txt'\n");
+		const std::string &s = buf.str();
+		fwrite(s.c_str(), 1, s.size(), f);
+		fclose(f);
+	}
 
 	initted = true;
